refactor(widgets): NotificationsWidget sous 300 lignes — éclatement complet (H8)
- composable useNotificationsWidget (état + chargement/marquage/suppression/navigation)
- sous-composant présentationnel NotificationItem (CSS verbatim, helpers type/icône)
- orchestrateur 552→285 lignes
- tests: composable (4) + NotificationItem (5)
- dette pré-existante conservée verbatim : styles .loading-state/.spinner morts (template emploie <ContentLoader>)
